refactor: move entity_relationship FQN hash migrations from v1.14.1 to v1.13.0
DDL (ADD COLUMN + INDEX for fromFQNHash/toFQNHash) and the Java backfill
(backfillRelationshipFqnHashes) both move into the v1.13.0 migration so
existing deployments upgrading to v1.13 get the columns and backfill in
one shot.

- v1.13.0 mysql/postgres schemaChanges.sql: append ADD COLUMN + CREATE INDEX
- v1.13.0 MigrationUtil: add backfillRelationshipFqnHashes + helpers
- v1.13.0 mysql/postgres Migration.java: call backfill after existing logic
- v1.14.1 schemaChanges.sql: no-op (columns already added by v1.13.0)
- v1.14.1 Migration.java: no-op (backfill already done by v1.13.0)
- Remove v1.14.1/MigrationUtil.java (logic merged into v1.13.0)
