+++ openmetadata-service/src/main/java/org/openmetadata/service/migration/mysql/v1141/Migration.java
@@ -13,29 +13,12 @@
 
 package org.openmetadata.service.migration.mysql.v1141;
 
-import lombok.SneakyThrows;
-import lombok.extern.slf4j.Slf4j;
 import org.openmetadata.service.migration.api.MigrationProcessImpl;
 import org.openmetadata.service.migration.utils.MigrationFile;
-import org.openmetadata.service.migration.utils.v1141.MigrationUtil;
 
-@Slf4j
 public class Migration extends MigrationProcessImpl {
 
   public Migration(MigrationFile migrationFile) {
     super(migrationFile);
-  }
-
-  @Override
-  @SneakyThrows
-  public void runDataMigration() {
-    try {
-      MigrationUtil.backfillRelationshipFqnHashes(handle);
-    } catch (Exception e) {
-      LOG.error(
-          "Failed to backfill FQN hashes in entity_relationship during v1141 migration. "
-              + "Fast prefix deletion may not work correctly for pre-existing relationships.",
-          e);
-    }
   }
 }
